api_server/requirements.txt
fastapi
uvicorn
slowapi
pandas
google-genai